almost authed, just need to add sign in form

+++ backend/src/resolvers/Mutation/post.js
@@ -1,8 +1,10 @@
-const { getUserId } = require('../../utils')
+const { getUserId } = require("../../utils");
 
 const post = {
   async createDraft(parent, { title, text }, ctx, info) {
-    const userId = getUserId(ctx)
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in to create a draft");
+    }
     return ctx.db.mutation.createPost(
       {
         data: {
@@ -10,45 +12,49 @@ const post = {
           text,
           isPublished: false,
           author: {
-            connect: { id: userId },
-          },
-        },
+            connect: { id: userId }
+          }
+        }
       },
       info
-    )
+    );
   },
 
   async publish(parent, { id }, ctx, info) {
-    const userId = getUserId(ctx)
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in to publish a post");
+    }
     const postExists = await ctx.db.exists.Post({
       id,
-      author: { id: userId },
-    })
+      author: { id: userId }
+    });
     if (!postExists) {
-      throw new Error(`Post not found or you're not the author`)
+      throw new Error(`Post not found or you're not the author`);
     }
 
     return ctx.db.mutation.updatePost(
       {
         where: { id },
-        data: { isPublished: true },
+        data: { isPublished: true }
       },
-      info,
-    )
+      info
+    );
   },
 
   async deletePost(parent, { id }, ctx, info) {
-    const userId = getUserId(ctx)
+    if (!ctx.request.userId) {
+      throw new Error("You must be logged in to delete a post");
+    }
     const postExists = await ctx.db.exists.Post({
       id,
-      author: { id: userId },
-    })
+      author: { id: userId }
+    });
     if (!postExists) {
-      throw new Error(`Post not found or you're not the author`)
+      throw new Error(`Post not found or you're not the author`);
     }
 
-    return ctx.db.mutation.deletePost({ where: { id } })
-  },
-}
+    return ctx.db.mutation.deletePost({ where: { id } });
+  }
+};
 
-module.exports = { post }
+module.exports = { post };
